unsigned na "população"

+++ CartasSuperTrunfo.c
@@ -5,7 +5,7 @@ int main() {
     char estado1;
     char codigo1[4];
     char cidade1[50];
-    int populacao1;
+    unsigned long int populacao1;
     float area1;
     float pib1;
     int pontosTuristicos1;
@@ -14,7 +14,7 @@ int main() {
     char estado2;
     char codigo2[4];
     char cidade2[50];
-    int populacao2;
+    unsigned long int populacao2;
     float area2;
     float pib2;
     int pontosTuristicos2;
